fix: unify demo helpers and restore tailwind base
