+++ src/index.css
@@ -4,7 +4,6 @@
 
 @layer base {
   :root {
-    /* Light theme (shadcn defaults) */
     --background: 0 0% 100%;
     --foreground: 222.2 84% 4.9%;
     --muted: 210 40% 96.1%;
@@ -27,7 +26,6 @@
     --radius: 0.5rem;
   }
   .dark {
-    /* Dark theme (shadcn defaults) */
     --background: 222.2 84% 4.9%;
     --foreground: 210 40% 98%;
     --muted: 217.2 32.6% 17.5%;
@@ -49,7 +47,6 @@
     --ring: 217.2 32.6% 17.5%;
   }
 
-  /* Base resets without @apply */
   html, body { height: 100%; }
   body { background-color: hsl(var(--background)); color: hsl(var(--foreground)); }
   *, ::before, ::after { border-color: hsl(var(--border)); }
